refactor op and raw_ops
