+++ src/eager.rs
@@ -10,4 +10,5 @@ mod tensor_handle;
 pub use tensor_handle::*;
 
 mod op;
-pub use op::*;
+
+pub mod raw_ops;
